move some common classes to bbb-common-message for spring integration changes
This commit simply moves a few of the classes to bbb-common-message and makes all of the related build changes to support this change.
No functionality was changed (at least intentionally).

Next step is to change the current Spring Integration messages (conference started / stopped) to send the whole conference in the message and confirm that this is working as intended.  Then it's easy to add other API calls